+++ src/main/java/org/devlogtwo/devlog/domain/task/controller/TaskController.java
@@ -1,17 +1,14 @@
 package org.devlogtwo.devlog.domain.task.controller;
 
 import jakarta.validation.Valid;
-import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import org.devlogtwo.devlog.common.code.SuccessCode;
 import org.devlogtwo.devlog.common.dto.GlobalApiResponse;
 import org.devlogtwo.devlog.common.util.ResponseHelper;
-import org.devlogtwo.devlog.domain.task.dto.TaskCreateRequest;
-import org.devlogtwo.devlog.domain.task.dto.TaskCreateResponse;
-import org.devlogtwo.devlog.domain.task.entity.Task;
+import org.devlogtwo.devlog.domain.task.dto.request.TaskCreateRequest;
+import org.devlogtwo.devlog.domain.task.dto.response.TaskCreateResponse;
 import org.devlogtwo.devlog.domain.task.service.TaskService;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
